+++ fabric/src/main/java/io/github/sakurawald/fuji/core/command/argument/adapter/impl/DimensionTypeArgumentTypeAdapter.java
@@ -1,13 +1,13 @@
 package io.github.sakurawald.fuji.core.command.argument.adapter.impl;
 
 import com.mojang.brigadier.arguments.ArgumentType;
-import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.RequiredArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import io.github.sakurawald.fuji.core.auxiliary.minecraft.CommandHelper;
 import io.github.sakurawald.fuji.core.command.argument.adapter.abst.BaseArgumentTypeAdapter;
 import io.github.sakurawald.fuji.core.command.argument.structure.Argument;
 import io.github.sakurawald.fuji.core.command.argument.wrapper.impl.DimensionType;
+import net.minecraft.command.argument.IdentifierArgumentType;
 import net.minecraft.registry.RegistryKeys;
 import net.minecraft.server.command.ServerCommandSource;
 
@@ -17,12 +17,12 @@ public class DimensionTypeArgumentTypeAdapter extends BaseArgumentTypeAdapter {
 
     @Override
     protected ArgumentType<?> makeArgumentType() {
-        return StringArgumentType.greedyString();
+        return IdentifierArgumentType.identifier();
     }
 
     @Override
     public Object makeArgumentObject(CommandContext<ServerCommandSource> context, Argument argument) {
-        return new DimensionType(StringArgumentType.getString(context, argument.getArgumentName()));
+        return new DimensionType(IdentifierArgumentType.getIdentifier(context, argument.getArgumentName()).toString());
     }
 
     @Override